refactor(durationfmt): extract plural rules to internal/ilib_plural
Move the CLDR plural-rule evaluation out of ILibDurationFmt into
stateless top-level functions (getPluralCategory) under internal/.
The logic depends only on the rule map, so a class added no value.
